f2c: merge libf2c recipe
