+++ recipes/f2c/all/patches/f2c/0002-prototype-rmdir.patch
