+++ recipes/f2c/all/patches/f2c/0004-fc-script-path-and-tmpfile.patch
