+++ recipes/f2c/all/patches/f2c/0008-typos.patch
