+++ recipes/f2c/all/patches/libf2c/0006-weak-MAIN__.patch
