+++ recipes/f2c/all/patches/libf2c/0009-F77_aloc-type.patch
