+++ recipes/f2c/all/patches/libf2c/0011-cross.patch
