+++ recipes/f2c/all/patches/libf2c/0012-enable-shared.patch
