--- recipes/libf2c/all/conandata.yml
@@ -1,20 +1,0 @@
-sources:
-  "20240130":
-    url:
-      - "https://github.com/valgur/f2c/releases/download/libf2c-20240130/libf2c-20240130.zip"
-      - "https://www.netlib.org/f2c/libf2c.zip"
-    sha256: "cc84253b47b5c036aa1d529332a6c218a39ff71c76974296262b03776f822695"
-patches:
-  # https://salsa.debian.org/debian/libf2c2/-/tree/7de354efec5312e15deab83affdaa620ae75849a/debian/patches
-  - patch_file: "patches/0003-C-tweaks.patch"
-  - patch_file: "patches/0004-add-clapack-files.patch"
-  - patch_file: "patches/0006-weak-MAIN__.patch"
-  - patch_file: "patches/0007-format-security-sprintf.patch"
-  - patch_file: "patches/0008-XintXX_t.patch"
-  - patch_file: "patches/0009-F77_aloc-type.patch"
-  - patch_file: "patches/0010-largefile_source-redefinition.patch"
-  - patch_file: "patches/0011-cross.patch"
-  # CCI
-  - patch_file: "patches/0012-enable-shared.patch"
-    patch_description: "Enable building of shared library"
-    patch_type: "conan"
