--- recipes/libf2c/config.yml
@@ -1,3 +1,0 @@
-versions:
-  "20240130":
-    folder: all
